requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.0.1
chess==1.9.4
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Pillow==9.4.0
yarl==1.8.2
